Confirmed C code works. Not interfacing with python
