+++ analysis/cython/install.sh
