+++ analysis/kmeans_C.py
@@ -7,52 +7,18 @@ import matplotlib
 from sklearn.decomposition import PCA
 import matplotlib.pyplot as plt
 from scipy.stats import gaussian_kde
-from kM import kmeansCPY as kM
+import kM
 
 #matplotlib inline
 
 
-mpl.rc("savefig", dpi=100)
-
-sns.set_style("white")
-
-greys = sns.color_palette("Greys", 10)
-
 data = pd.read_csv('../data/yeast.csv', header=None).values
 #k-means witk sklearn
-kmeans = KMeans(n_clusters=3,n_init=100)#,algorithm='full')
-kmeans.fit(data)
-
-from sklearn.decomposition import PCA
-import matplotlib.pyplot as plt
-
-pca = PCA(n_components=2)#, svd_solver='full')
-pca.fit(data)
-
-pcs   = pca.transform(data)
-cntrs = pca.transform(kmeans.cluster_centers_)
-
-mycmap=matplotlib.colors.ListedColormap(greys[3:])
-
-x,y=pcs[:,0],pcs[:,1]
-
-# Calculate the point density
-xy = np.vstack([x,y])
-z = gaussian_kde(xy)(xy)
-
-plt.figure(figsize=(6,6),dpi=100)
-plt.scatter(x, y, c=z, s=30, edgecolor='white',cmap=mycmap)
-plt.scatter(cntrs[:,0],cntrs[:,1],facecolor='red',edgecolor='black',s=120,alpha=0.95)
-plt.xlabel('PC1')
-plt.xlabel('PC2')
-
-#score
-kmeans.inertia_
 
 K=3
 N,D=data.shape
 
-A = np.zeros((K,D))
+A = np.zeros((K,D), dtype=np.float64)
 W = np.zeros(N,dtype=np.intc)
 X = data
 m = np.zeros(K)
@@ -114,38 +80,10 @@ while not converged:
             W[n] = min_ind
             converged=False
 
-pca = PCA(n_components=2)#, svd_solver='full')
-pca.fit(data)
-
-pcs   = pca.transform(data)
-cntrs = pca.transform(A)
-
-mycmap=matplotlib.colors.ListedColormap(greys[3:])
-
-x,y=pcs[:,0],pcs[:,1]
-
-# Calculate the point density
-xy = np.vstack([x,y])
-z = gaussian_kde(xy)(xy)
-
-plt.figure(figsize=(6.5,6.5))
-plt.scatter(x, y, c=z, s=30, edgecolor='white',cmap=mycmap)
-plt.scatter(cntrs[:,0],cntrs[:,1],facecolor='red',edgecolor='black',s=120,alpha=0.95)
-plt.xlabel('PC1')
-plt.xlabel('PC2')
-
-phi=0
-
-for n in range(N):
-    for d in range(D):
-        phi += (X[n,d]-A[W[n],d])**2
-
-
-print(phi)
 
 c = np.zeros((K,D))
 x = data
-WW = np.zeros(N,dtype=np.intc)
+WW = np.zeros(N,dtype=np.int32)
 
 def shuffle(x,n):
     for i in range(n-2,-1,-1): #from n-2 to 0
@@ -166,9 +104,8 @@ x = x.copy(order='C')
 X = X.copy(order='C')
 A = A.copy(order='C')
 
-kM.SS(X, A)
 
-kM.kM(x, c, WW)
+kM.kMeans(x, c, WW)
 
 print("Stock K-mean equal to Eric's clusters?  ", end=" ")
 print(np.array_equal(kmeans.cluster_centers_,c))
@@ -187,3 +124,7 @@ print("Kareem v. Eric Clusters", end=" ")
 print(np.array_equal(c,A))
 
 exit()
+
+
+gcc -std=c11 -c src/kmeans.c kmeans_wrap.c \
+    -I/usr/local/include/python2.1
